cloud/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
redis>=5.0.0
celery>=5.3.0
httpx>=0.27.0
pyjwt>=2.8.0
cryptography>=42.0.0
websockets>=12.0
python-multipart>=0.0.9
